+++ src/DataSet.h
@@ -82,7 +82,7 @@ class TabReader {
   int readInt() {
     char* next = NULL;
     errno=0;
-    int i = strtol(f_, &next, 10);
+    int i = static_cast<int>(strtol(f_, &next, 10));
     if (next == f_) {
       err = 1;
     } else {
@@ -126,7 +126,7 @@ class DataSet {
   void inline setLabel(int l) { label_ = l; }
   int inline getLabel() const { return label_; }
   
-  unsigned int inline getSize() const { return psms_.size(); }
+  unsigned int inline getSize() const { return static_cast<unsigned int>(psms_.size()); }
   
   static inline void setCalcDoc(bool on) { calcDOC_ = on; }
   static inline bool getCalcDoc() { return calcDOC_; }
@@ -136,7 +136,7 @@ class DataSet {
     featureNames_ = FeatureNames();
     FeatureNames::resetNumFeatures();
   }
-  static unsigned getNumFeatures() { return featureNames_.getNumFeatures(); }
+  static unsigned getNumFeatures() { return static_cast<unsigned>(featureNames_.getNumFeatures()); }
   
   bool writeTabData(std::ofstream& out);
